pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/steklov_arap"]

[project]
name = "steklov-arap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cholespy>=2.2.0",
    "libigl>=2.6.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "polyscope>=2.6.1",
    "pyglet<2",
    "robust-laplacian>=1.0.0",
    "scipy>=1.17.1",
    "steklovnet",
    "torch>=2.10.0",
    "viser>=1.0.24",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "k3d>=2.17.0",
    "marimo>=0.23.3",
    "pytest>=9.0.2",
]

# [[tool.uv.index]]
# name = "pytorch-cu130"
# url = "https://download.pytorch.org/whl/cu130"
# explicit = true

[tool.uv.sources]
# torch = { index = "pytorch-cu130" }
steklovnet = { git = "https://github.com/starboi-63/steklovnet.git", rev = "installable" }
